럍: 전투 오류 수정
플레이어 공격, 적 방어 시 적의 받는 피해 경감이 제대로 이뤄지지 않던 문제 수정

+++ src/Character.cpp
@@ -27,7 +27,7 @@ void Character::setHp(int atk, int def) {
     }
 
     //방어율과 실제 입는 피해 계산 후 체력 차감
-    double reduction = def / 100.0;       
+    double reduction = (double)def / 100.0;       
     double damage = atk * (1.0 - reduction);
     this->hp -= static_cast<int>(damage);
 
@@ -35,8 +35,6 @@ void Character::setHp(int atk, int def) {
     if (this->hp < 0) {
         this->hp = 0;
     }
-
-    
 }
